Move LightDarkMode enums into the Logic file and refactored.
